requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
boto3==1.28.62
python-dotenv==1.0.0 